Move across "figures of merit"
