+++ cfspopcon/formulas/fusion_power/fusion_gain.py
@@ -31,3 +31,18 @@ def calc_fusion_gain(P_fusion: Unitfull, P_alpha: Unitfull, P_in: Unitfull, frac
     Q = Q.where(Q < _IGNITED_THRESHOLD, _IGNITED)
 
     return P_external, P_launched, Q
+
+
+@Algorithm.register_algorithm(return_keys=["fusion_triple_product"])
+def calc_triple_product(peak_fuel_ion_density: Unitfull, peak_ion_temp: Unitfull, energy_confinement_time: Unitfull) -> Unitfull:
+    """Calculate the fusion triple product.
+
+    Args:
+        peak_fuel_ion_density: [1e20 m^-3] :term:`glossary link<peak_fuel_ion_density>`
+        peak_ion_temp: [keV] :term:`glossary link<peak_fuel_ion_density>`
+        energy_confinement_time: [s] :term:`glossary link<energy_confinement_time>`
+
+    Returns:
+         fusion_triple_product [10e20 m**-3 keV s]
+    """
+    return peak_fuel_ion_density * peak_ion_temp * energy_confinement_time
