+++ cfspopcon/formulas/metrics/__init__.py
@@ -1,5 +1,5 @@
 """Various metrics for operating points, generally used when we can't fit something into another logical grouping."""
 
-from . import greenwald_density, heat_exhaust_metrics
+from . import collisionality, greenwald_density, heat_exhaust_metrics, rho_star
 
-__all__ = ["heat_exhaust_metrics", "greenwald_density"]
+__all__ = ["heat_exhaust_metrics", "greenwald_density", "collisionality", "rho_star"]
